use MuPDF's mutool for PDF file information
Support mutool as an alternative to pdftk for
- determining page count and presentation title
- uncompressing PDF files before parsing them for hyperlinks

pdftk is now only required for extracting page titles from bookmarks.
For everything else, mutool is sufficient (and, in case of uncompressing,
even superior to pdftk).

As a result, Impressive is now reasonably useable with only a single
external program: mutool(.exe) from MuPDF >= 1.9 (which is the first version
that drops mudraw in favor of "mutool draw").


git-svn-id: https://svn.emphy.de/impressive/trunk/impressive@194 58003caa-664d-4c56-a378-6d187ff66ee1