major new feature: box-zoom mode
Ctrl+LMB-drag now enters a special zoom mode
where the selected rectangle is zoomed in to
cover the whole screen, and the rest of the
screen is considerably dimmed down (even more
so than in highlight box and spotlight mode).

This also decouples the visual zoom factor
from the page resolution zoom factor;
in other words, it's possible to zoom in
*really* far now, even though the rendering
resolution is limited to only 5x, 2x or
even 1x (see new --maxzoom option).


git-svn-id: https://svn.emphy.de/impressive/trunk/impressive@218 58003caa-664d-4c56-a378-6d187ff66ee1