rudimentary support for TransitionTo calls in PageProps
Trying to call TransitionTo from a lambda in a 'OnEnter' or
'OnLeave' PageProps didn't work properly before. Now, it's
somewhat possible, but in the case of 'OnLeave(Once)', if
and only if the new keyword argument notify_page_left=False
is set.

TL;DR: Info scripts like this work now:
    PageProps = {
       10: { 'OnLeaveOnce':
             lambda: TransitionTo(5, notify_page_left=False)
           }
    }


git-svn-id: https://svn.emphy.de/impressive/trunk/impressive@292 58003caa-664d-4c56-a378-6d187ff66ee1