improved pip-local SDL library path detection
Some pip (or pygame?) versions don't put the binary libraries
in .../site-packages/pygame/.libs, but .../site-packages/pygame.libs
(note the missing slash!). This change accounts for that.

I didn't manage to find any documentation or specification
with a definitive answer to where those libraries are
supposed to be located, so we'll keep this heuristic for now.


git-svn-id: https://svn.emphy.de/impressive/trunk/impressive@305 58003caa-664d-4c56-a378-6d187ff66ee1