more overview mode event handling fixes
Overview mode keyboard navigation relied on the fact
that setting the mouse position works and generates
a mouse motion event. However, in PyGame 2.0 / SDL 2.0,
this is no longer the case if the mouse cursor wasn't
already inside the window.

This also contains a *proper* fix (and partial revert)
for the issue in r308.


git-svn-id: https://svn.emphy.de/impressive/trunk/impressive@309 58003caa-664d-4c56-a378-6d187ff66ee1